Module pallet_paged_list::mock

Helpers for tests.

Structs

MaxPages

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

RuntimeGenesisConfig

RuntimeOrigin

The runtime origin type representing the origin of a call.

Test

ValuesPerNewPage

Enums

OriginCaller

RuntimeCall

The aggregated runtime call type.

RuntimeError

RuntimeEvent

RuntimeFreezeReason

A reason for placing a freeze on funds.

RuntimeHoldReason

A reason for placing a hold on funds.

RuntimeLockId

An identifier for each lock placed on funds.

RuntimeSlashReason

A reason for slashing funds.

RuntimeTask

An aggregation of all Task enums across all pallets included in the current runtime.

Functions

new_test_ext

Build genesis storage according to the mock runtime.

test_closure

Run this closure in test externalities.

Type Aliases

AllPalletsWithSystem

All pallets included in the runtime as a nested tuple of types.

AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

MetaOf

PagedList

PagedList2

System

SystemConfig