Module pallet_paged_list::mock

Helpers for tests.

Structs

MaxPages

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

RuntimeGenesisConfig

RuntimeOrigin

The runtime origin type representing the origin of a call.

Test

ValuesPerNewPage

Enums

OriginCaller

RuntimeCall

RuntimeError

RuntimeEvent

RuntimeFreezeReason

A reason for placing a freeze on funds.

RuntimeHoldReason

A reason for placing a hold on funds.

RuntimeLockId

An identifier for each lock placed on funds.

RuntimeSlashReason

A reason for slashing funds.

Functions

new_test_ext

Build genesis storage according to the mock runtime.

test_closure

Run this closure in test externalities.

Type Definitions

AllPalletsDeprecated

All pallets included in the runtime as a nested tuple of types.

AllPalletsReversedWithSystemFirstDeprecated

All pallets included in the runtime as a nested tuple of types in reversed order. With the system pallet first.

AllPalletsWithSystem

All pallets included in the runtime as a nested tuple of types.

AllPalletsWithSystemReversedDeprecated

All pallets included in the runtime as a nested tuple of types in reversed order.

AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

AllPalletsWithoutSystemReversedDeprecated

All pallets included in the runtime as a nested tuple of types in reversed order. Excludes the System pallet.

GenesisConfigDeprecated

MetaOf

PagedList

PagedList2

System

SystemConfig